Consolidate all database info into DatabaseInfo
