+++ src/test/java/comp4111/AuthenticationIntegrationTest.java
@@ -16,6 +16,7 @@ import org.junit.jupiter.api.Test;
 
 import java.sql.DriverManager;
 
+import static comp4111.dal.DatabaseInfo.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
@@ -29,7 +30,7 @@ public class AuthenticationIntegrationTest extends AbstractServerTest {
         super.setUp();
 
         assumeTrue(() -> {
-            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(DatabaseConnection.MYSQL_URL, DatabaseConnection.MYSQL_LOGIN, DatabaseConnection.MYSQL_PASSWORD)) {
+            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(MYSQL_URL, MYSQL_LOGIN, MYSQL_PASSWORD)) {
                 return true;
             } catch (Throwable tr) {
                 return false;
@@ -152,7 +153,7 @@ public class AuthenticationIntegrationTest extends AbstractServerTest {
         }
         super.tearDown();
 
-        DatabaseUtils.dropDatabase(DatabaseConnection.DB_NAME);
+        DatabaseUtils.dropDatabase(DB_NAME);
         DatabaseConnectionPoolV2.getInstance().close();
         DatabaseConnection.cleanUp();
