+++ src/test/java/comp4111/BookOperationsAndTransactionIntegrationTest.java
@@ -19,6 +19,7 @@ import org.junit.jupiter.api.Test;
 import java.sql.DriverManager;
 import java.util.regex.Pattern;
 
+import static comp4111.dal.DatabaseInfo.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
@@ -37,7 +38,7 @@ public class BookOperationsAndTransactionIntegrationTest extends AbstractServerT
         super.setUp();
 
         assumeTrue(() -> {
-            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(DatabaseConnection.MYSQL_URL, DatabaseConnection.MYSQL_LOGIN, DatabaseConnection.MYSQL_PASSWORD)) {
+            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(MYSQL_URL, MYSQL_LOGIN, MYSQL_PASSWORD)) {
                 return true;
             } catch (Throwable tr) {
                 return false;
@@ -453,7 +454,7 @@ public class BookOperationsAndTransactionIntegrationTest extends AbstractServerT
         }
         super.tearDown();
 
-        DatabaseUtils.dropDatabase(DatabaseConnection.DB_NAME);
+        DatabaseUtils.dropDatabase(DB_NAME);
         DatabaseConnectionPoolV2.getInstance().close();
         DatabaseConnection.cleanUp();
