+++ src/test/java/comp4111/bug/Bug9_GetBooksSmallerThanLimitTest.java
@@ -22,6 +22,7 @@ import org.junit.jupiter.api.Test;
 import java.sql.DriverManager;
 import java.util.regex.Pattern;
 
+import static comp4111.dal.DatabaseInfo.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
@@ -36,7 +37,7 @@ public class Bug9_GetBooksSmallerThanLimitTest extends AbstractServerTest {
         super.setUp();
 
         assumeTrue(() -> {
-            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(DatabaseConnection.MYSQL_URL, DatabaseConnection.MYSQL_LOGIN, DatabaseConnection.MYSQL_PASSWORD)) {
+            try (@SuppressWarnings("unused") var con = DriverManager.getConnection(MYSQL_URL, MYSQL_LOGIN, MYSQL_PASSWORD)) {
                 return true;
             } catch (Throwable tr) {
                 return false;
@@ -165,7 +166,7 @@ public class Bug9_GetBooksSmallerThanLimitTest extends AbstractServerTest {
         }
         super.tearDown();
 
-        DatabaseUtils.dropDatabase(DatabaseConnection.DB_NAME);
+        DatabaseUtils.dropDatabase(DB_NAME);
         DatabaseConnection.cleanUp();
 
         token = null;
